code/frontend/requirements.txt
flask==2.0.1
requests==2.26.0
google-auth==2.3.0
Werkzeug==2.3.7